The install command should reference a RPM filename comprised both of the module's name and version.
This ensures that the glob only picks up what should be installed and nothing more.

This manifested on an attempted install of Fey. Fey-Loader had already existed in the build directory, so when Fey attempted to install, the glob also attempted to install Fey-Loader again which generated an already-installed error and caused Fey to (silently) not be installed.